Extract dataProvider in simple example for easier testing
