+++ examples/simple/dataProvider.js
@@ -1,0 +1,16 @@
+import jsonRestDataProvider from 'ra-data-fakerest';
+
+import data from './data';
+import addUploadFeature from './addUploadFeature';
+
+const dataProvider = jsonRestDataProvider(data, true);
+const uploadCapableDataProvider = addUploadFeature(dataProvider);
+const delayedDataProvider = (type, resource, params) =>
+    new Promise(resolve =>
+        setTimeout(
+            () => resolve(uploadCapableDataProvider(type, resource, params)),
+            1000
+        )
+    );
+
+export default delayedDataProvider;
